fastapi-backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
alembic==1.12.1
email-validator==2.1.0
python-dotenv==1.0.0


httpx==0.27.0
authlib==1.2.0
itsdangerous==2.1.2